application/backend/pyproject.toml
[project]
name = "physical-ai-studio"
version = "1.0.0"
description = "Physical AI Studio server"
requires-python = ">=3.12"

dependencies = [
  "opencv-python",
  "cv2-enumerate-cameras",
  'harvesters; sys_platform != "darwin"',
  "fastapi[standard]",
  "physicalai-train[pi0,smolvla]",
  "lerobot[feetech]==0.4.4",
  "numpy",
  "websockets",
  'pyrealsense2-macosx; sys_platform == "darwin"',
  'pyrealsense2; sys_platform != "darwin"',
  "aiortc>=1.13.0",
  "pydantic-settings>=2.10.1",
  "sqlalchemy>=2.0.43",
  "aiosqlite~=0.21",
  "alembic>=1.16.5",
  "click>=8.3.0",
  "FrameSource>=0.2.2",
  "loguru>=0.7.3",
  "greenlet>=3.2.4",
  "tenacity>=9.1.2",
  "aiofiles>=25.1.0",
  "types-aiofiles>=25.1.0.20251011",
  "trossen-arm==1.9.0",
  "pyzmq>=27.1.0",
  "sse-starlette>=2.0.0"
]

[project.optional-dependencies]
cpu = [
    "torch<2.11",
    "torchvision<0.26.0",
]
cuda = [
    "torch<2.11",
    "torchvision<0.26.0",
]
xpu = [
    "torch<2.11",
    "torchvision<0.26.0",
    "pytorch-triton-xpu ; sys_platform == 'linux' or sys_platform == 'win32'",
]

tests = [
    "pytest",
]

# --- PyTorch index configuration ---
# torch is a transitive dependency (via lightning, lerobot) that defaults to
# PyPI's fat CUDA wheel on Linux. Source routing here directs uv to fetch
# torch/torchvision from the correct hardware-specific PyTorch index.
# This must be declared in the consumer project because uv does not
# propagate [tool.uv.sources] from path dependencies (uv #14675).
#
# PyPI is listed first as a named default index so that it takes priority
# over the non-explicit pytorch-xpu index in unsafe-best-match resolution.
# Without this, the XPU index can shadow PyPI packages with incompatible
# wheels (e.g. markupsafe cp314-only, see uv #9647).
[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[[tool.uv.index]]
name = "pytorch-xpu"
url = "https://download.pytorch.org/whl/xpu"
explicit = false

[tool.uv.sources]
FrameSource = { git = "https://github.com/ArendJanKramer/FrameSource.git", rev = "c3023714ceaa2bf50cf04e1d2861d9fe7cd01422" }
physicalai-train = { path = "../../library", editable = true }
torch = [
    { index = "pytorch-cpu", extra = "cpu" },
    { index = "pytorch-cu128", extra = "cuda" },
    { index = "pytorch-xpu", extra = "xpu" },
]
torchvision = [
    { index = "pytorch-cpu", extra = "cpu" },
    { index = "pytorch-cu128", extra = "cuda" },
    { index = "pytorch-xpu", extra = "xpu" },
]
pytorch-triton-xpu = [{ index = "pytorch-xpu", extra = "xpu" }]

[dependency-groups]
dev = [
  "ruff~=0.11.2",
  "pyrefly>=0.18.0",
]
lint = [
  "ruff~=0.11.2",
  "mypy~=1.17",
  "pyrefly>=0.18.0",
]
all = [
  "ruff~=0.11.2",
  "mypy~=1.17",
  "pyrefly>=0.18.0",
]

[tool.uv]
default-groups = ["all"]
# The pytorch-xpu index is non-explicit (explicit = false) because
# transitive deps like pytorch-triton-xpu only exist on the XPU index.
# unsafe-best-match is required because the XPU index also hosts stale
# copies of common packages (e.g. cmake 3.25.0); with first-match uv
# would stop at the XPU index and miss the newer PyPI version.
#
# To prevent the XPU index from providing incompatible wheels for
# common packages (e.g. markupsafe cp314-only, see uv #9647), PyPI
# is listed as a named index with higher priority above.
index-strategy = "unsafe-best-match"
conflicts = [
    [
        { extra = "cpu" },
        { extra = "cuda" },
        { extra = "xpu" },
    ],
]

# ============================================================================ #
# RUFF CONFIGURATION                                                           #
# ============================================================================ #
[tool.ruff]
# Extend shared configuration from root
extend = "../../pyproject.toml"

# Disable preview mode for backend (inherited as true from root)
preview = false

# Backend-specific source directories
src = ["src"]

# Backend-specific excludes
extend-exclude = [".venv*"]

[tool.ruff.lint]
# Backend uses a subset of rules (different from library's ALL)
select = [
    "ARG", "E", "F", "I", "N", "UP", "YTT", "ASYNC", "S", "COM", "C4", "FA",
    "PIE", "PYI", "Q", "RSE", "RET", "SIM", "TID", "TC", "PL", "RUF", "C90",
    "D103", "ANN001", "ANN201", "ANN205", "FAST"
]

ignore = [
    "N801", "N805", "N806", "N807", "N818", "COM812", "RET503", "SIM108",
    "SIM105", "PLR2004", "RUF010", "TC001", "RUF012"
]

fixable = ["ALL"]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
"*test*.py" = ["S", "ANN", "D", "ARG", "PLR"]

[tool.ruff.lint.isort]
split-on-trailing-comma = false

[tool.ruff.lint.pylint]
max-args = 7


# ============================================================================ #
# MYPY CONFIGURATION                                                           #
# ============================================================================ #
[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
show_error_codes = true

[tool.pyrefly]
python-version = "3.12"
search-path = ["src"]
project-includes = ["src"]
project-excludes = [
  "src/alembic/versions/**",
  ".venv/**",
  "dist/**",
]

[tool.pytest.ini_options]
addopts = ["--strict-markers", "--strict-config", "--showlocals", "-ra"]
testpaths = "tests"
pythonpath = "src"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "requires_download: marks tests that require downloading external datasets",
    "integration: marks tests as integration tests",
]
filterwarnings = [
  "ignore:XPU device count is zero!:UserWarning",
  "ignore:This process.*is multi-threaded:DeprecationWarning", # using actual queues in tests
]
